Updated render_nlp_tags and autogen.py
